refactor(netlist): extract the IR model into netlist_models.py
First cut splitting the 6,325-line netlist god-module by responsibility. The
25 model dataclasses/enums (NetRef … NetlistModule, the NetlistItem alias, the
build context) move to a new netlist_models.py (748 lines); netlist.py
re-exports every name via `from .netlist_models import (...)` so all existing
`from ...graph.netlist import <Name>` call sites are unchanged. netlist.py
6,325 → 5,635.

Behavior-preserving: lvnet output byte-identical, `pyright src/` clean, the
netlist/lvnet test subset green (full suite runs in the pre-commit hook). No
import cycle (netlist_models never imports netlist).
